+++ flutter_inappwebview_forge_platform_interface/lib/src/types/permission_request.g.dart
@@ -42,13 +42,16 @@ class PermissionRequest {
     );
     if (map['resources'] != null) {
       instance.resources = List<PermissionResourceType>.from(
-        map['resources'].map(
-          (e) => switch (enumMethod ?? EnumMethod.nativeValue) {
-            EnumMethod.nativeValue => PermissionResourceType.fromNativeValue(e),
-            EnumMethod.value => PermissionResourceType.fromValue(e),
-            EnumMethod.name => PermissionResourceType.byName(e),
-          }!,
-        ),
+        map['resources']
+            .map(
+              (e) => switch (enumMethod ?? EnumMethod.nativeValue) {
+                EnumMethod.nativeValue =>
+                  PermissionResourceType.fromNativeValue(e),
+                EnumMethod.value => PermissionResourceType.fromValue(e),
+                EnumMethod.name => PermissionResourceType.byName(e),
+              },
+            )
+            .whereType<PermissionResourceType>(),
       );
     }
     return instance;