+++ flutter_inappwebview_forge_windows/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1 - 2026-08-06
+
+- Prevent forward-incompatible WebView2 permission resource values from crashing permission request decoding; known resources remain available to applications.
+
 ## 1.0.0
 
 - First `flutter_inappwebview_forge_windows` release as part of the Forge federated plugin.